+++ WalletWasabi/Crypto/ZeroKnowledge/LinearRelation/Knowledge.cs
@@ -4,12 +4,10 @@ using WalletWasabi.Helpers;
 
 namespace WalletWasabi.Crypto.ZeroKnowledge.LinearRelation
 {
-	public class Knowledge
+	internal class Knowledge
 	{
-		public Knowledge(Statement statement, ScalarVector witness)
+		internal Knowledge(Statement statement, ScalarVector witness)
 		{
-			Guard.NotNull(nameof(statement), statement);
-			Guard.NotNullOrEmpty(nameof(witness), witness);
 			Guard.True(nameof(witness), witness.Count() == statement.Equations.First().Generators.Count(), $"{nameof(witness)} size does not match {nameof(statement)}.{nameof(statement.Equations)}");
 
 			// don't try to prove something which isn't true
